Move `dandi organize` code to an `organize()` function
